requirements.txt
certifi==2018.8.24
chardet==3.0.4
feedparser==5.2.1
idna==2.7
lxml==4.2.5
python-dateutil==2.7.3
python-vlc==3.0.102
requests==2.19.1
six==1.11.0
urllib3==1.23
watson-developer-cloud==2.0.1
weather-api==1.0.4
websocket-client==0.47.0
